Colors and fonts and borders (oh my!)
I'm not a designer. This is a joke. Whoever was like "Hey Carter,
you should be in charge of some colors and making things look good"
was mental. I used Comic Sans. Non-Ironically. Jesus. I should
never be in charge of making things that are even remotely user-facing.
Please, please, please let me spend the rest of my life making really
sick products with nothing more than a CLI.

In this commit:
- changing the font of the buttons to comic sans
- Making the colors look slightly less like vomit
- Rounding the borders of my button divs (oh my)
- Adding borders to the blocks (the color is firebrick. I chose it because
    the name is dope)